requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
yt-dlp>=2024.0.0
douyin-tiktok-scraper
edge-tts>=6.1.0
pytest>=8.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
